Nicer physics
Physics have been adjusted. Gravity is set way higher, because Unity
always assumes object sizes are in meters. Otherwise we'd have to make the
scale really small for decent jumping, which the editor doesn't like.

Removed the need for platforms to have names/tags, because they're
annoying to assign.

Small debug level added with actual jumps.

Camera moved to the right to show more of terrain about to be encountered,
instead of centered between ahead & behind