Added in the falling/exploding platform functionality
How to use it:
1.  Place a Destructible Platform Trigger in the Scene -- make sure to
set a platform to it.
2.  Place a Destructible Platform & set the following values
a.  Force To Apply -- Direction of the Force
b.  Point to Apply Force -- Point about which the force is applied
(Torque)
c.  Force Magnitude -- How Strong the Initial Force is
d.  Constant Vector Force -- Force to apply after the initial force
e.  How long to apply the Constant Vector Force for.