Grayboxed the Second part of level 1
1.  Red Areas are blocks that can have forces applied to when a trigger
is hit.
2.  Red Areas have the following properties:
a.  The properties from the last commit
b.  Delay Time -- when should the box have forces acted on it?
c.  Post Activation Constraints -- when a block is activated, what
constraints are acted upon it?
Triggers can now hold arrays of destructible platforms, which can be
used to create much more complex behavior.