Added in base explosion support to all lasers and missiles for L1
How it works right now is that every laser and missile has a default
detonator prefab, which can be customized.

When you want to create a laser from a laser trigger, just set
IsDetonator and the default will be used.

If you want to use a special detonator from the laser trigger for a
particular laser, just set that is the postCollisionParticle system.