I think I finally fixed the shield bug.
It was going on in the Reset Shield Method of the L2 ship, due to the
special case of hitting the elite ship only, which is why the bug never
came up in L4.  It was an issue with the coroutines being called and
threading issues with just using a single bool.